Remove redundant Gemini ApiComplianceTest and consolidate into per-concern files
