+++ tests/Feature/Providers/Gemini/ToolMappingTest.php
@@ -7,32 +7,6 @@ use Tests\Feature\Agents\ToolUsingAgent;
 
 class ToolMappingTest extends GeminiTestCase
 {
-    public function test_tool_parameters_use_object_type(): void
-    {
-        Http::fake([
-            'generativelanguage.googleapis.com/*' => $this->fakeTextResponse('The number is 42'),
-        ]);
-
-        (new ToolUsingAgent(fixed: true))->prompt(
-            'Generate a number',
-            provider: 'gemini',
-        );
-
-        Http::assertSent(function ($request) {
-            $tools = $request->data()['tools'] ?? [];
-
-            foreach ($tools as $toolGroup) {
-                foreach ($toolGroup['function_declarations'] ?? [] as $decl) {
-                    if ($decl['name'] === 'FixedNumberGenerator') {
-                        return ! isset($decl['parameters']);
-                    }
-                }
-            }
-
-            return false;
-        });
-    }
-
     public function test_empty_schema_omits_parameters_key(): void
     {
         Http::fake([
@@ -100,8 +74,15 @@ class ToolMappingTest extends GeminiTestCase
         Http::assertSent(function ($request) {
             $tools = $request->data()['tools'] ?? [];
 
-            return isset($tools[0]['function_declarations'])
-                && count($tools[0]['function_declarations']) > 0;
+            if (! isset($tools[0]['function_declarations']) || count($tools[0]['function_declarations']) === 0) {
+                return false;
+            }
+
+            $decl = $tools[0]['function_declarations'][0];
+
+            return isset($decl['name'])
+                && isset($decl['description'])
+                && is_string($decl['description']);
         });
     }
 }
